requirements.txt
requests>=2.31.0  # For API calls
python-dotenv>=1.0.0  # Optional: For .env file support